Magnus Lind ( magli143 ) - 2012-07-14 10:48:49 PDT
This patch doubles the fragment size if the sound mode is stereo compared to
when it is mono since the double amount of data is written per sample in stereo
mode. Without this patch, when using stereo mode, the size of the sound buffer
(in milliseconds) is only half the configured size.

http://sourceforge.net/tracker/?func=detail&atid=1057619&aid=3544080&group_id=223021


git-svn-id: svn://svn.code.sf.net/p/vice-emu/code/trunk@25979 379a1393-f5fb-40a0-bcee-ef074d9b53f7